Fix for reading offset and looking up buffer at that offset without validating offset > buffer.length. This causes input[i] below being undefined in each such case.
if (input[i] !== 0xff) {

This is important when using imageSize() with Uint8Array as buffer and reading the file in chuked mode, up until it actually returns the dimensions.
The following fix will cause TypeError to be thrown earlier, and now users of imageSize() will know they need to read in another chunk from the file,
and call the function again, on the new, larger buffer.