feat: Load full local image file when required
Adds a fallback strategy to `imageSizeFromFile` in case that the image size information is
not contained in the first 512kb of the files read initially, and is known to exist outside
(implemented only for `.tiff` files for now): The whole file is read from disk, and the
`imageSize` function called again with the full file content.

This solves cases where - for some `.tiff` files, getting the image size would throw an
error instead of returning the image size, based on the image size information not being
contained in the first 512kb.

A custom error class is used as interface to trigger the fallback.