Make SVG validation more robust
In image-size/image-size#375, SVG validation was changed to searching
for the <svg ...> tag in the first 1000 bytes. This is quite brittle and
yields false negatives for draw.io SVG exports, whose <svg> tag is very
long. Instead of limiting length and requiring a terminal '>', just
search the Uint8Array for '<svg' without string conversion.